[#44] Nurse can book an appointment for Patient
Slight bug in this; if any change is made in one tab and the other one was loaded before, the pre-loaded one will not reflect changes as the state is of the component is not updated. For instance, if 'Appointments' tab is opened, then a booking is made, when revisiting 'Appointments' tab, the new appointment will not be there as the component is not updated. To see the new update in the 'Appointments' tab, the page will have to be reloaded.